+++ vulture_ignorelist.py
@@ -1,1 +1,2 @@
-format_suffix  # unused variable (codex/views/reader/page.py:41)
+_PermissionClass  # unused import (codex/views/admin/auth.py:13)
+format_suffix  # unused variable (codex/views/opds/binary.py:27)